+++ lib/utilities/hourly_temp.dart
@@ -5,14 +5,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:http/http.dart' as http;
+import 'package:provider/provider.dart';
+import 'package:weather_app/providers/weather_data_provider.dart';
 import 'helper_functions.dart';
-
-class HourlyTempData {
-  String temp;
-  String hour;
-
-  HourlyTempData({required this.temp, required this.hour});
-}
 
 // Component displays hourly temprature for a day
 class HourlyTemp extends StatelessWidget {
@@ -59,70 +54,11 @@ class HourlyTempAll extends StatefulWidget {
 }
 
 class _HourlyTempAllState extends State<HourlyTempAll> {
-  late List<HourlyTempData> temps;
-  bool _dataFetched = false; // Flag to track whether data has been fetched
-
-  @override
-  void initState() {
-    super.initState();
-    temps = []; // Initialize temps list
-    _fetchHourlyTemp();
-  }
-
-  Future<void> _fetchHourlyTemp() async {
-    final queryParams = {
-      'lat': widget.lat,
-      'lon': widget.long,
-      'units': "metric",
-      'appid': dotenv.env["WEATHER_API_KEY"],
-    };
-
-    final uri =
-        Uri.parse("https://pro.openweathermap.org/data/2.5/forecast/hourly")
-            .replace(queryParameters: queryParams);
-
-    final response = await http.get(uri);
-
-    if (response.statusCode == 200) {
-      final jsonData = jsonDecode(response.body);
-
-      final List<dynamic> hourlyData = jsonData['list'];
-      List<HourlyTempData> tempList = [];
-
-      for (int i = 0; i < 24; i++) {
-        HourlyTempData entry = HourlyTempData(
-          temp: hourlyData[i]["main"]["temp"].round().toString(),
-          hour: formatTimestamp(hourlyData[i]["dt"]),
-        );
-        tempList.add(entry);
-      }
-      print("Hourly");
-      setState(() {
-        temps = tempList;
-        _dataFetched = true;
-      });
-    } else {
-      throw Exception('Failed to load hourly temperature');
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
-    return _dataFetched
-        ? _HourlyTempList(temps: temps)
-        : Center(
-            child: CircularProgressIndicator(),
-          );
-  }
-}
+    print("Hourly");
+    final providerWeather = Provider.of<WeatherDataProvider>(context);
 
-class _HourlyTempList extends StatelessWidget {
-  final List<HourlyTempData> temps;
-
-  const _HourlyTempList({Key? key, required this.temps}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
     return SizedBox(
       height: 180,
       width: 380,
@@ -131,15 +67,15 @@ class _HourlyTempList extends StatelessWidget {
         child: Row(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            for (int i = 0; i < temps.length; i++)
+            for (int i = 0; i < providerWeather.tempList.length; i++)
               Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 10.0),
+                padding: EdgeInsets.symmetric(horizontal: 10.0),
                 child: Container(
                   clipBehavior: Clip.none,
                   width: 100, // Adjust width as needed
                   child: HourlyTemp(
-                    time: temps[i].hour,
-                    temp: '${temps[i].temp}°C',
+                    time: providerWeather.tempList[i].hour,
+                    temp: '${providerWeather.tempList[i].temp}°C',
                   ),
                 ),
               ),
